requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart==0.0.6
google-genai>=1.56.0
Pillow>=10.1.0,<12.0.0
python-dotenv==1.0.0
openai>=1.68.0
